+++ src/main/java/com/gagoo/thiscoding/domain/maria/user/service/MyPageServiceImpl.java
@@ -3,8 +3,8 @@ package com.gagoo.thiscoding.domain.maria.user.service;
 import com.gagoo.thiscoding.domain.maria.user.controller.port.MyPageService;
 import com.gagoo.thiscoding.domain.maria.user.domain.User;
 import com.gagoo.thiscoding.domain.mongo.board.service.port.BoardCustomRepository;
-import com.gagoo.thiscoding.global.security.SecurityUtils;
 import com.gagoo.thiscoding.domain.maria.user.service.port.UserRepository;
+import com.gagoo.thiscoding.global.security.service.port.SecurityUtils;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -16,10 +16,11 @@ public class MyPageServiceImpl implements MyPageService {
 
     private final UserRepository userRepository;
     private final BoardCustomRepository boardCustomRepository;
+    private final SecurityUtils securityUtils;
 
     @Override
     public Boolean isTop10() {
-        User currentUser = userRepository.getByEmail(SecurityUtils.getUserEmail());
+        User currentUser = userRepository.getByEmail(securityUtils.getUserEmail());
 
         List<Long> top10Users = boardCustomRepository.getTop10Users();
         boolean isTop10 = validateTop10(currentUser, top10Users);
