+++ src/main/java/com/gagoo/thiscoding/global/security/service/port/SecurityUtils.java
@@ -1,0 +1,6 @@
+package com.gagoo.thiscoding.global.security.service.port;
+
+public interface SecurityUtils {
+
+    String getUserEmail();
+}
